pyproject.toml
[tool.poetry]
name = "pocarisweat"
version = "0.1.0"
description = ""
authors = ["Ryota Fujimura <ryotaajishio0626@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
antlr4-python3-runtime = "4.8"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
hydra-core = "1.1.2"
idna = "3.3"
importlib-resources = "5.2.3"
joblib = "1.1.0"
numpy = "1.23.5"
omegaconf = "2.1.2"
pandas = "1.4.2"
pillow = "9.1.0"
python-dateutil = "2.8.2"
pytz = "2022.1"
pyyaml = "6.0"
requests = "2.27.1"
scikit-learn = "1.0.2"
scipy = "1.8.0"
six = "1.16.0"
sklearn = "0.0"
threadpoolctl = "3.1.0"
torch = "1.11.0"
torchvision = "0.12.0"
tqdm = "4.64.0"
typing-extensions = "4.2.0"
urllib3 = "1.26.9"
zipp = "3.8.0"
optuna = "^3.2.0"
xgboost = "^2.0.2"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow = "^2.10.0"
tensorflow_text = "2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
